requirements/test.txt
# Test dependencies go here.
-r base.txt
--extra-index-url https://test.pypi.org/simple

coverage==4.4.1
flake8==3.4.1 # pyup: != 2.6.0
django-test-plus==1.0.18
factory-boy==2.9.2
django-coverage-plugin==1.5.0

# pytest
pytest-django==3.1.2
# temporarily disabled
# pytest-sugar==0.9.0

# django-courier
git+https://gitlab.com/empurrandojuntos/django-courier.git
